Create MostOccuringValue
You are given an array of integers that contain numbers in random order. Write a program to find and return the number which occurs the maximum times in the given input. If two or more elements contend for the maximum frequency, return the element which occurs in the array first.